Move the cache and update references
